+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 See also [https://github.com/stkb/vscode-rewrap/releases](https://github.com/stkb/vscode-rewrap/releases) (for working links to issues)
 
 
+## Unreleased
+
+- Support CJK content (#75)
+
+
 ### 1.8.1
 
 - Support Vue HTML templates (#82)